refactor: relocate onboarding schemas, unify form hook, add explicit abort, and add submit hook tests (Xchange-Taiwan/X-Talent-Tracker#344)
